+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## v0.2.0 (2025-01-29)
+
+- Changed the way how secondary services for smart home devices are added, because sometimes HomeKit doesn't play nice
+- Added optional characteristics for lightbulbs (brightness, color temperature, hue/saturation). Don't know if this works as expected. I have no devices to test.
+- Added support for HAN-FUN devices. Might work. I have no devices to test.
+
 ## v0.1.0 (2025-01-26)
 
 Initial commit